+++ tests/get_assertion.rs
@@ -1,5 +1,5 @@
 fn test<'data, T: serde::Deserialize<'data> + std::fmt::Debug>(data: &'data [u8]) {
-    let result = ctap_types::serde::cbor_deserialize::<T>(data);
+    let result = cbor_smol::cbor_deserialize::<T>(data);
     assert!(result.is_ok(), "{:?}", result);
 }