+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [Unreleased]: https://github.com/trussed-dev/ctap-types/compare/0.6.0-rc.4...HEAD
 
 - Remove unused `Rpc` trait.
+- Remove `cbor-smol` re-export as `serde`.
 
 ## [0.6.0-rc.4] 2026-06-01